managed ei - refactoring
